edit: 수업일지 API 변경 사항 반영
