+++ src/page/community/ui/Community.tsx
@@ -1,9 +1,6 @@
 'use client';
 
-import Link from 'next/link';
-
 import { useMyInfoQuery } from '@/feature/member';
-import { IconBack } from '@/shared/assets';
 import { Typography } from '@/shared/mixin';
 import { cn } from '@/shared/utils';
 import { Layout } from '@/widget';
@@ -13,9 +10,6 @@ const Community = () => {
   return (
     <Layout type='trainer'>
       <Layout.Header className='relative'>
-        <Link href='/trainer/manage'>
-          <IconBack />
-        </Link>
         <h1
           className={cn(
             Typography.HEADING_4_SEMIBOLD,
